workaround to support static and server build
