+++ geoda-ai/next.config.js
@@ -13,17 +13,6 @@ const nextConfig = {
   reactStrictMode: false,
   // enable a static export and export all pages to static HTML files under out folder
   output: isStaticExport ? 'export' : 'standalone',
-  // exportPathMap: async function (defaultPathMap) {
-  //   if (isStaticExport) {
-  //     // Example: Delete specific dynamic API route entries
-  //     Object.keys(defaultPathMap).forEach(path => {
-  //       if (path.startsWith('/api/auth/')) {
-  //         delete defaultPathMap[path];
-  //       }
-  //     });
-  //   }
-  //   return defaultPathMap;
-  // },
   // discard: only set basePath to /reactgeoda to deploy in github pages
   basePath: process.env.BASE_PATH ?? '/reactgeoda',
   typescript: {
@@ -125,14 +114,6 @@ const nextConfig = {
         replace: keplerVersion,
         flags: 'g'
       }
-    });
-
-    config.module.rules.push({
-      test: filePath => {
-        // Check if the file path includes the /app/api directory
-        return filePath.includes(path.resolve(__dirname, 'api', 'auth'));
-      },
-      use: 'null-loader'
     });
 
     // Define global variable __VERSION__ with value "4.1.0" for loaders.gl
