+++ geoda-ai/src/app/providers.tsx
@@ -1,10 +1,13 @@
-'use client';
 import React from 'react';
 import {NextUIProvider} from '@nextui-org/react';
 import {SessionProvider} from 'next-auth/react';
 
+const IS_STATIC_EXPORT = process.env.NEXT_PUBLIC_STATIC_EXPORT === 'true';
+
 export function Providers({children}: {children: React.ReactNode}) {
-  return (
+  return IS_STATIC_EXPORT ? (
+    <NextUIProvider>{children}</NextUIProvider>
+  ) : (
     <SessionProvider>
       <NextUIProvider>{children}</NextUIProvider>
     </SessionProvider>
